major: moved CBREngine.java from no.ntnu.mycbr to no.ntnu.mycbr.rest.cbr package
